### 2026-01-26 #9906 @editorialbot
[PRE REVIEW]: biostats: Biostatistics and Clinical Data Analysis in R
**Submitting author:** @sebasquirarte (<a href="http://orcid.org/0009-0004-2730-4343">Sebastian Quirarte-Justo</a>)
**Repository:** https://github.com/sebasquirarte/biostats/
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.1.1
