### 2026-01-27 76e4cc2
Removed examples section from JOSS paper
