+++ paper.md
@@ -235,223 +235,6 @@ labeling to enhance interpretability. Each function returns a fully customizable
 ggplot2 object, allowing users to refine themes, annotations, scales, and other 
 graphical elements.
 
-# Examples
-
-``` r
-# Simulate basic clinical data
-clinical_df <- clinical_data()
-
-head(clinical_df, 5)
-#>    participant_id visit  sex treatment age weight biomarker response
-#> 1             001     1 Male Treatment  35   55.4     42.22 Complete
-#> 2             001     2 Male Treatment  35   60.3     44.70     None
-#> 3             001     3 Male Treatment  35   58.1     44.85  Partial
-#> 4             002     1 Male   Placebo  21   68.3     56.51     None
-#> 5             002     2 Male   Placebo  21   66.3     51.03     None
-#> 6             002     3 Male   Placebo  21   64.0     39.59     None
-```
-
-``` r
-# Grouped summary by treatment group with all stats and effect size
-summary_table(clinical_df, group_by = 'treatment', all = TRUE,
-              effect_size = TRUE, exclude = c('participant_id', 'visit'))
-```
-
-![](figures/figure3.png)
-
-``` r
-# Filter clinical data to Placebo arm
-clinical_df_treat <- clinical_df[clinical_df$treatment == "Placebo", ]
-
-# Normally distributed variable
-normality(data = clinical_df_treat, "biomarker")
-#> 
-#> Normality Test for 'biomarker' 
-#> 
-#> n = 159 
-#> mean (SD) = 49.44 (9.2) 
-#> median (IQR) = 50.38 (13.1) 
-#> 
-#> Kolmogorov-Smirnov (Lilliefors): D = 0.054, p = 0.305 
-#> Shapiro-Wilk: W = 0.992, p = 0.546 
-#> Skewness: 0.06 (z = 0.30) 
-#> Kurtosis: -0.03 (z = -0.08) 
-#> 
-#> Data appears normally distributed.
-```
-
-![](figures/figure4.png){ width=85% }
-
-``` r
-# Simulate more complex clinical data
-clinical_df_full <- clinical_data(n = 300, visits = 10, arms = c('A', 'B', 'C'), 
-                                  dropout = 0.10, missing = 0.05)
-                                  
-# Missing value analysis of only variables with missing values
-missing_values(clinical_df_full)
-#> 
-#> Missing Value Analysis
-#> 
-#> Complete rows: 2452 (81.7%)
-#> Missing cells: 868 (3.6%)
-#> 
-#>           n_missing pct_missing
-#> response        403       13.43
-#> weight          251        8.37
-#> biomarker       214        7.13
-```
-
-![](figures/figure5.png){ width=75% }
-
-``` r
-# Basic outlier detection
-outliers(clinical_df_full, "biomarker")
-#> 
-#> Outlier Analysis
-#> 
-#> Variable: 'biomarker'
-#> n: 2786
-#> Missing: 214 (7.1%)
-#> Method: Tukey's IQR x 1.5
-#> Bounds: [18.971, 74.761]
-#> Outliers detected: 19 (0.7%)
-#> 
-#> Outlier indices: 27, 223, 440, 559, 795, 931, 973, 1175, 1277, 1346 (...)
-```
-
-![](figures/figure6.png){ width=75% }
-
-``` r
-# Two-sample parallel non-inferiority test for means with 10% expected dropout
-sample_size(sample = 'two-sample', design = 'parallel', outcome = 'mean',
-            type = 'non-inferiority', x1 = 5.0, x2 = 5.0, 
-            SD = 0.1, delta = -0.05, k = 1, dropout = 0.1)
-#> 
[…]
