+++ paper.bib
@@ -5,16 +5,17 @@
   address = {Vienna, Austria},
   year = {2024},
   url = {https://www.R-project.org/},
+  doi = {10.32614/r.manuals}
 }
 
 @article{Sagar:2023,
-  title   = {Biostatistical Methodologies in Clinical Trials: An Overview of Recent Developments and Pitfalls},
-  author  = {Sagar, K. P. and P., Rachna and B., Sandeep},
+  title = {Biostatistical Methodologies in Clinical Trials: An Overview of Recent Developments and Pitfalls},
+  author = {Sagar, K. P. and P., Rachna and B., Sandeep},
   journal = {International Journal of Clinical Biostatistics and Biometrics},
-  volume  = {9},
-  pages   = {051},
-  year    = {2023},
-  doi     = {10.23937/2469-5831/1510051}
+  volume = {9},
+  pages = {051},
+  year = {2023},
+  doi = {10.23937/2469-5831/1510051}
 }
 
 @article{Ciolino:2021,
@@ -29,14 +30,14 @@
 }
 
 @article{Dwivedi:2022,
-  title   = {How to write statistical analysis section in medical research},
-  author  = {Dwivedi, A. K.},
+  title = {How to write statistical analysis section in medical research},
+  author = {Dwivedi, A. K.},
   journal = {Journal of Investigative Medicine},
-  volume  = {70},
-  number  = {8},
-  pages   = {1759--1770},
-  year    = {2022},
-  doi     = {10.1136/jim-2022-002479}
+  volume = {70},
+  number = {8},
+  pages = {1759--1770},
+  year = {2022},
+  doi = {10.1136/jim-2022-002479}
 }
 
 @Manual{Hmisc:2026,
@@ -45,6 +46,7 @@
   year = {2026},
   note = {R package version 5.2-5},
   url = {https://CRAN.R-project.org/package=Hmisc},
+  doi = {10.32614/cran.package.hmisc}
 }
 
 @Manual{tableone:2022,
@@ -54,6 +56,7 @@ without Propensity Score Weights},
   year = {2022},
   note = {R package version 0.13.2},
   url = {https://CRAN.R-project.org/package=tableone},
+  doi = {10.32614/cran.package.tableone}
 }
 
 @Manual{pwr:2020,
@@ -62,6 +65,7 @@ without Propensity Score Weights},
   year = {2020},
   note = {R package version 1.3-0},
   url = {https://CRAN.R-project.org/package=pwr},
+  doi = {10.32614/cran.package.pwr}
 }
 
 @Article{effectsize:2020,
@@ -92,6 +96,7 @@ without Propensity Score Weights},
   year = {2016},
   note = {R package version 4.4-0},
   url = {https://CRAN.R-project.org/package=ez},
+  doi = {10.32614/cran.package.ez}
 }
 
 @Manual{rstatix:2023,
@@ -100,6 +105,7 @@ without Propensity Score Weights},
   year = {2023},
   note = {R package version 0.7.2},
   url = {https://CRAN.R-project.org/package=rstatix},
+  doi = {10.32614/cran.package.rstatix}
 }
 
 @Manual{ggblanket:2025,
@@ -108,6 +114,7 @@ without Propensity Score Weights},
   year = {2025},
   note = {R package version 12.4.0},
   url = {https://CRAN.R-project.org/package=ggblanket},
+  doi = {10.32614/cran.package.ggblanket}
 }
 
 @Manual{ggpubr:2025,
@@ -116,6 +123,7 @@ without Propensity Score Weights},
   year = {2025},
   note = {R package version 0.6.1},
   url = {https://CRAN.R-project.org/package=ggpubr},
[…]
