### 2026-01-27 #9906 @sebasquirarte
> [@sebasquirarte](https://github.com/sebasquirarte) - Thank you for your submission to JOSS. We are working through a queue of submissions and will get you a topic editor assigned as soon as one becomes available.
>
> In the meantime, please remove the "examples" section from your paper to help us get down to a desired word count of around 1500 words max. You can simply reference your documentation if you want to point to examples from the paper.
>
> Also, please check the missing DOI listed [here](https://github.com/openjournals/joss-reviews/issues/9906#issuecomment-3802218214).

Thank you for the follow-up, I have removed the 'Examples' section and added the missing DOIs as suggested.
