### 2026-01-31 #9906 @ajayi-joseph
Hi, I'd be happy to review this submission once an editor is assigned. I have years of experience in web development with JavaScript and have built serverless applications. My background includes production systems in healthcare and fintech.
