### 2026-03-05 2503287
Bump version to 1.1.2 for CRAN submission
