+++ cran-comments.md
@@ -1,20 +1,16 @@
-## biostats 1.1.0 – CRAN submission
+## biostats 1.1.2 – CRAN submission
 
 ### R CMD check results
 0 errors | 0 warnings | 0 notes
 
-### Changes in this version (1.1.0)
+### Changes in this version (1.1.2)
 
 #### New Features
-* Implemented Lilliefors correction for Kolmogorov–Smirnov normality testing in both `normality()` and `summary_table()` functions. This provides more accurate p-values when normal distribution parameters are estimated from sample data.
-
-#### Improvements
-* Improved p-value formatting to use `"p < 0.001"` instead of `"p = < 0.001"`.
-* Added methodological references: Lilliefors (1967) and Dallal & Wilkinson (1986).
-* Updated documentation to clarify that the K-S option now uses the Lilliefors correction.
-
-#### Dependencies
-* Added `nortest` to Imports.
+* Fixed dropout adjustment formula in `sample_size()`.
+  The previous implementation used `n * (1 + dropout)`, a first-order
+  approximation that underestimates required enrollment at higher dropout
+  rates. The correct formula `n / (1 - dropout)` is now used, consistent
+  with Chow et al. (2017) and ICH E9 guidelines.
 
 ### Notes to CRAN
-* This is a minor update adding Lilliefors correction functionality and documentation refinements. No breaking changes were introduced.
+* This is a minor update fixing the dropout formula in the sample_size() function. No breaking changes were introduced.
