+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: biostats
 Title: Biostatistics and Clinical Data Analysis
-Version: 1.1.1
+Version: 1.1.2
 Authors@R: c(
     person("Sebastian", "Quirarte-Justo", 
            email = "sebastian.quirarte@sophia.com.mx", 
