### 2026-03-16 #9906 @Nikoleta-v3
Hey everyone! Sorry, that really took me a moment to recover 😅

In the meantime, I had some time to look over the submission. Here are a few comments @sebasquirarte.

### Paper

Typos:

- 	Line 62: unifies these analysis → unifies these analyses
- 	Line 67: an clear → a clear

Length:

I do find the paper a bit on the lengthy side. The examples section is quite long, but I would prefer to hear the reviewers’ perspective on this as well.

###  README

On a related note, the README is also quite long. The issue is that as a user scrolls to a section, there is no easy way to navigate back apart from scrolling all the way up. I would recommend having dedicated online documentation instead. One simple solution would be to use the GitHub Wiki:
https://docs.github.com/en/communities/documenting-your-project-with-wikis/about-wikis (but it's a paid feature).

However, a nicer approach might be something like `pkgdown` which generates a full documentation website for R packages and can be deployed automatically using GitHub Actions and GitHub Pages. For example, another submission I recently reviewed has documentation like this:
https://github.com/jianhuig/fairmetrics.

Alternatively, you could use Sphinx, which is widely used for Python projects:
https://www.sphinx-doc.org.

I think users would benefit from having such dedicated documentation rather than a very long `README`. Of course, I realize this is not trivial to set up, but I would be happy to support you in doing so if helpful.

### Reviewers

I would also like to start inviting reviewers for the submission soon. Have you had a chance to look over the reviewer database and suggest a few possible reviewers? https://reviewers.joss.theoj.org/lookup
