+++ paper.md
@@ -106,14 +106,14 @@ assessment via distinct tests (e.g. Shapiro–Wilk, Kolmogorov–Smirnov), kurto
 measures, or independent graphical analyses; sample size calculations through 
 functions tailored to individual study designs; and effect measures evaluated 
 separately for each type of association. In contrast, the _**biostats**_ package 
-unifies these analysis within single functions, providing a unified, consistent,
+unifies these analyses within single functions, providing a unified, consistent,
 and streamlined workflow.
 
 The *omnibus()* function offers an integrated approach to determining whether 
 parametric linear models or non-parametric alternatives are appropriate. It 
 evaluates data using minimally specified parameters, returns the corresponding 
 model’s analysis, reports observed values per each assumption, runs appropriate 
-post-hoc tests, and  presents the results in an clear and easy-to-follow format.  
+post-hoc tests, and  presents the results in adee clear and easy-to-follow format.  
 
 The *missing_values()*, *outliers()*, and *summary_table()* functions present data and 
 analysis in a clean and organized format with professional visual outputs, as 
