### 2026-03-16 #9906 @sebasquirarte
Hi @Nikoleta-v3, thank you for the follow-up. Regarding your comments and suggestions:

- I have updated the joss-paper branch to be aligned with the main branch and latest version of the package released on CRAN (v.1.1.2). The typos mentioned have also been fixed.

- As suggested by @crvernon as well, the examples section has been removed from the paper all together. Please let me know if the updated version of the paper is not available.

- I will work on implementing `pkgdown` to move the documentation to GitHub pages and keep the repository README short and easier to navigate. Can we proceed with the review process while this is being worked on or is it necessary to implement before moving forward?

- In terms of potential reviewers, @ajayi-joseph I appreciate you reaching out and expressing your interest in participating, if possible I would be happy to have you as a reviewer. Other potential reviewers I have found on the database that I think could be interested in this project are Salvador Balkus (salbalkus) and/or Avi Kenny (Avi-Kenny).

Please let me know if I should suggest additional reviewers or if you need anything from my end.
