### 2026-03-17 #9906 @Nikoleta-v3
Thank you, @sebasquirarte! As you can see from the compiled pdf, the changes have been applied. However, the typo in line  `67`was not fixed 😅  it now says `adee`.

Thank you as well for agreeing to work on the documentation; I think it will benefit the package a lot.

I will go ahead and start working on finding a second reviewer for the submission, but I would like the documentation to be completed before the official review begins so that the reviewers can also review the online documentation and the updated README.
