### 2026-03-25 #9906 @Nikoleta-v3
@AmaanArif25 @rachelicr 👋🏻 Would any of you be willing to review this submission for JOSS?

The submission is titled: “biostats: Biostatistics and Clinical Data Analysis in R". You can find more details at the top of the GitHub issue for the submission. 🆙

As you already know, we conduct our checklist-driven reviews here in GitHub issues and follow these guidelines:
https://joss.readthedocs.io/en/latest/review_criteria.html 😄
