### 2026-03-25 #9906 @chrisarg
I am in the middle of travel and a grant review for our center so I have to
pass on this one.

Christos

On Fri, Mar 20, 2026, 12:40 AM Nikoleta Glynatsi ***@***.***>
wrote:

> *Nikoleta-v3* left a comment (openjournals/joss-reviews#9906)
> <https://github.com/openjournals/joss-reviews/issues/9906#issuecomment-4095978248>
>
> @smwu <https://github.com/smwu> @chrisarg <https://github.com/chrisarg>👋🏻
> Would any of you be willing to review this submission for JOSS?
>
> The submission is titled: “biostats: Biostatistics and Clinical Data
> Analysis in R". You can find more details at the top of the GitHub issue
> for the submission. 🆙
>
> As you already know, we conduct our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html 😄
>
