### 2026-03-31 6a83783
Add structured pkgdown site with Get Started vignette and shortened README
