+++ README.Rmd
@@ -20,6 +20,7 @@ knitr::opts_chunk$set(collapse = TRUE,
 ![CRAN status](https://www.r-pkg.org/badges/version/biostats)
 [![R-CMD-check](https://github.com/sebasquirarte/biostats/actions/workflows/R-CMD-check.yml/badge.svg)](https://github.com/sebasquirarte/biostats/actions/workflows/R-CMD-check.yml)
 [![codecov](https://codecov.io/gh/sebasquirarte/biostats/branch/main/graph/badge.svg)](https://app.codecov.io/gh/sebasquirarte/biostats)
+[![status](https://joss.theoj.org/papers/bb1fafdfe2c83e989a3f1578d2e50dc1/status.svg)](https://joss.theoj.org/papers/bb1fafdfe2c83e989a3f1578d2e50dc1)
 [![Total Downloads](https://cranlogs.r-pkg.org/badges/grand-total/biostats)](https://cranlogs.r-pkg.org/)
 
 <!-- badges: end -->
@@ -58,549 +59,6 @@ pak::pak("sebasquirarte/biostats")
 
 ```{r usage, warning=FALSE, message=FALSE, results='hide'}
 library(biostats)
-```
-
-This package comprises 14 functions across four analytical domains: 
-
-- [**Descriptive Statistics and Exploratory Data Analysis (EDA)**](#descriptive-statistics-and-exploratory-data-analysis-eda)
-  - [clinical_data()](#clinical_data)
-  - [summary_table()](#summary_table)
-  - [normality()](#normality)
-  - [missing_values()](#missing_values) 
-  - [outliers()](#outliers)
-- [**Sample Size and Power Calculation**](#sample-size-and-power-calculation)
-  - [sample_size()](#sample_size)
-  - [sample_size_range()](#sample_size_range)
-- [**Statistical Analysis and Inference**](#statistical-analysis-and-inference)
-  - [omnibus()](#omnibus) 
-  - [effect_measures()](#effect_measures) 
-- [**Data Visualization**](#data-visualization)
-  - [plot_bar()](#plot_bar)
-  - [plot_line()](#plot_line)
-  - [plot_hist()](#plot_hist)
-  - [plot_box()](#plot_box)
-  - [plot_corr()](#plot_corr)
-
-### Descriptive Statistics and Exploratory Data Analysis (EDA)
-
-#### **clinical_data()**
-
-##### Description
-
-Creates a simple simulated clinical trial dataset with subject demographics, multiple visits, treatment groups with different effects, numerical and categorical variables, as well as optional missing data and dropout rates.
-
-##### Parameters
-
-| Parameter | Description | Default |
-|-----------|-------------|---------|
-| `n` | Integer indicating the number (1-999) of subjects. | `100` |
-| `visits` | Integer indicating the number of visits including baseline. | `3` |
-| `arms` | Character vector of treatment arm names. | `c("Placebo", "Treatment")` |
-| `dropout` | Numeric parameter indicating the proportion (0-1) of subjects who dropout. | `0` |
-| `missing` | Numeric parameter indicating the proportion (0-1) of missing values to be introduced across numeric variables with fixed proportions (biomarker = 15%, weight = 25%, response = 60%). | `0` |
-
-
-##### Examples
-
-```{r}
-# Simulate basic clinical data
-clinical_df <- clinical_data()
-
-str(clinical_df)
-
-head(clinical_df, 10)
-```
-```{r}
-# Simulate more complex clinical data
-clinical_df_full <- clinical_data(n = 300,
-                                  visits = 10,
-                                  arms = c('A', 'B', 'C'), 
-                                  dropout = 0.10,
-                                  missing = 0.05)
-
[…]
