+++ vignettes/biostats.Rmd
@@ -1,0 +1,981 @@
+---
+title: "Get Started with biostats"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Get Started with biostats}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+This package comprises 14 functions across four analytical domains:
+
+- [**Descriptive Statistics and Exploratory Data Analysis
+  (EDA)**](#descriptive-statistics-and-exploratory-data-analysis-eda)
+  - [clinical_data()](#clinical_data)
+  - [summary_table()](#summary_table)
+  - [normality()](#normality)
+  - [missing_values()](#missing_values)
+  - [outliers()](#outliers)
+- [**Sample Size and Power
+  Calculation**](#sample-size-and-power-calculation)
+  - [sample_size()](#sample_size)
+  - [sample_size_range()](#sample_size_range)
+- [**Statistical Analysis and
+  Inference**](#statistical-analysis-and-inference)
+  - [omnibus()](#omnibus)
+  - [effect_measures()](#effect_measures)
+- [**Data Visualization**](#data-visualization)
+  - [plot_bar()](#plot_bar)
+  - [plot_line()](#plot_line)
+  - [plot_hist()](#plot_hist)
+  - [plot_box()](#plot_box)
+  - [plot_corr()](#plot_corr)
+
+To get started, load the biostats package:
+```{r setup}
+library(biostats)
+```
+
+### Descriptive Statistics and Exploratory Data Analysis (EDA)
+
+#### **clinical_data()**
+
+##### Description
+
+Creates a simple simulated clinical trial dataset with subject
+demographics, multiple visits, treatment groups with different effects,
+numerical and categorical variables, as well as optional missing data
+and dropout rates.
+
+##### Parameters
+
+| Parameter | Description | Default |
+|----|----|----|
+| `n` | Integer indicating the number (1-999) of subjects. | `100` |
+| `visits` | Integer indicating the number of visits including baseline. | `3` |
+| `arms` | Character vector of treatment arm names. | `c("Placebo", "Treatment")` |
+| `dropout` | Numeric parameter indicating the proportion (0-1) of subjects who dropout. | `0` |
+| `missing` | Numeric parameter indicating the proportion (0-1) of missing values to be introduced across numeric variables with fixed proportions (biomarker = 15%, weight = 25%, response = 60%). | `0` |
+
+##### Examples
+
+``` r
+# Simulate basic clinical data
+clinical_df <- clinical_data()
+
+str(clinical_df)
+#> 'data.frame':    300 obs. of  8 variables:
+#>  $ participant_id: chr  "001" "001" "001" "002" ...
+#>  $ visit         : Factor w/ 3 levels "1","2","3": 1 2 3 1 2 3 1 2 3 1 ...
+#>  $ sex           : Factor w/ 2 levels "Male","Female": 1 1 1 1 1 1 1 1 1 1 ...
+#>  $ treatment     : Factor w/ 2 levels "Placebo","Treatment": 2 2 2 1 1 1 1 1 1 1 ...
+#>  $ age           : num  35 35 35 21 21 21 47 47 47 35 ...
+#>  $ weight        : num  55.4 60.3 58.1 68.3 66.3 64 76 77.6 74.9 61.7 ...
+#>  $ biomarker     : num  42.2 44.7 44.9 56.5 51 ...
[…]
