+++ paper.md
@@ -1,0 +1,262 @@
+# Summary
+
+![](figures/logo.png)
+
+***biostats*** is an R package \[@R:2024\] that provides a cohesive and
+structured set of tools for biostatistics and clinical data analysis.
+The package includes 14 specialized functions covering descriptive
+statistics, exploratory data analysis, sample size and power
+calculations, statistical analysis and inference, and data
+visualization. These functions aim to offer standardized,
+well-documented workflows that are frequently required in clinical
+studies, trial planning, and analysis. By consolidating these
+capabilities into a single framework, the package facilitates
+consistent, transparent, and reproducible analyses across studies.
+
+This package serves both as an analytical toolkit for professional
+biostatisticians and clinical data analysts, and as an educational
+resource for researchers transitioning to R-based biostatistics,
+including professionals from other domains, clinical researchers, and
+medical practitioners involved in the development of clinical trials.
+
+***biostats*** is available on the Comprehensive R Archive Network
+(CRAN) and adheres to CRAN standards for documentation, testing,
+reproducibility, and long-term maintainability within the R ecosystem.
+
+![Functions included in the biostats package.](figures/figure1.png)
+
+Functions included in the biostats package.
+
+# Statement of need
+
+Biostatistics is a fundamental component of clinical research, essential
+for validating trial designs, methodologies, results, conclusions, as
+well as supporting submission to regulatory entities \[@Sagar:2023;
+@Ciolino:2021; @Dwivedi:2022\]. In practice, clinical data analysis
+involves the execution of similar tasks across multiple studies and
+projects. Typical workflows include the calculation of descriptive
+statistics and exploratory data analysis, assumption validation,
+hypothesis testing, primary, secondary, and exploratory statistical
+analyses, effect size estimation, sample size and power calculations, as
+well as data visualization.
+
+Popular packages in this field include Hmisc \[@Hmisc:2026\] and
+tableone \[@tableone:2022\] for descriptive statistics, pwr
+\[@pwr:2020\] for power and sample size calculations, effectsize
+\[@effectsize:2020\] for effect size estimates, and ggplot2
+\[@ggplot2:2016\] for data visualization, among others. While these
+packages are well-designed and widely used, completing a clinical study
+workflow typically requires combining multiple packages with different
+syntax conventions , output formats, and integration patterns. As a
+result, analysts frequently develop custom code to connect results,
+automate recurring tasks, or standardize outputs across studies. This
+fragmentation can lead to inconsistent implementations, duplicated
+effort, and increased time spent on code development, validation, and
+quality control.
+
+The ***biostats*** package addresses these challenges by providing a
+unified, clinically oriented framework that consolidates commonly used
+biostatistical procedures into a single, coherent toolkit. While users
+still retain full flexibility to write custom code tailored to
+study-specific needs, ***biostats*** is designed to streamline
+repetitive and foundational tasks in biostatistics and clinical data
+analysis through consistent syntax, harmonized outputs, and functions
+that reflect standard clinical workflows. Its goal is to deliver a
+professional-grade toolset for biostatisticians and clinical researchers
+while remaining accessible to data analysts from other fields. In
+addition, ***biostats*** serves as an educational resource for users
+transitioning to R or to biostatistics, offering a structured and
+reproducible approach aligned with contemporary recommendations for
+transparent and rigorous statistical practice.
+
+# State of the field
+
+Regarding the specific functions in this package, ***biostats*** differs
+from existing packages such as ez \[@ez:2016\], rstatix
+\[@rstatix:2023\], ggblanket \[@ggblanket:2025\], ggpubr
+\[@ggpubr:2025\], extras \[@extras:2025\], SampleSize4ClinicalTrials
+\[@SampleSize4ClinicalTrials:2021\], TrialSize \[@TrialSize:2024\],
[…]
