+++ reference/index.md
@@ -1,0 +1,35 @@
+# Package index
+
+## All functions
+
+- [`clinical_data()`](clinical_data.md) : Simulate Simple Clinical Trial
+  Data
+- [`effect_measures()`](effect_measures.md)
+  [`print(`*`<effect_measures>`*`)`](effect_measures.md) : Effect
+  Measures
+- [`missing_values()`](missing_values.md)
+  [`print(`*`<missing_values>`*`)`](missing_values.md)
+  [`print(`*`<outliers>`*`)`](missing_values.md) : Descriptive and
+  Visual Missing Value Assessment
+- [`normality()`](normality.md)
+  [`print(`*`<normality>`*`)`](normality.md) : Statistical and Visual
+  Normality Assessment
+- [`omnibus()`](omnibus.md) [`print(`*`<omnibus>`*`)`](omnibus.md) :
+  Omnibus Tests for Comparing Three or More Groups
+- [`outliers()`](outliers.md) : Descriptive and Visual Outlier
+  Assessment
+- [`plot_bar()`](plot_bar.md) : Create Simple Professional Bar Plots
+- [`plot_box()`](plot_box.md) : Create Simple Professional Box Plots
+- [`plot_corr()`](plot_corr.md) : Create Simple Professional Correlation
+  Matrix Plots
+- [`plot_hist()`](plot_hist.md) : Create Simple Professional Histogram
+  Plots
+- [`plot_line()`](plot_line.md) : Create Simple Professional Line Plots
+- [`sample_size()`](sample_size.md)
+  [`print(`*`<sample_size>`*`)`](sample_size.md) : Sample Size
+  Calculation for Clinical Trials
+- [`sample_size_range()`](sample_size_range.md)
+  [`print(`*`<sample_size_range>`*`)`](sample_size_range.md) : Calculate
+  and visualize sample size across a range of treatment effects
+- [`summary_table()`](summary_table.md) : Summary Table with Optional
+  Group Comparisons

[…]
