### 2026-03-31 #10317 @sebasquirarte
Hi @Nikoleta-v3, apologies for the delayed response, I was out of town and just returned.

Following up on the comments from the pre-review thread, I have fixed the typo on line 67, shortened the repository's README to be more concise, and moved the detailed documentation and examples to a dedicated pkgdown site (available [here](https://sebasquirarte.github.io/biostats/)).

Thank you @ajayi-joseph, @rachelicr, and @AmaanArif25 for accepting the invitation and taking the time to review this package. I look forward to your feedback and will address any questions or suggestions as quickly as possible.
