+++ pkgdown.yml
@@ -3,4 +3,4 @@ pkgdown: 2.2.0
 pkgdown_sha: ~
 articles:
   biostats: biostats.html
-last_built: 2026-03-31T18:52Z
+last_built: 2026-03-31T21:05Z
