+++ README.Rmd
@@ -14,7 +14,7 @@ knitr::opts_chunk$set(collapse = TRUE,
                       out.width = "100%")
 ```
 
-# biostats <a href="https://github.com/sebasquirarte/biostats/blob/main/man/figures/logo.png"><img src="man/figures/logo.png" align="right" height="138" alt="biostats_logo" /></a>
+# biostats <a href="https://sebasquirarte.github.io/biostats/"><img src="man/figures/logo.png" align="right" height="120" alt="biostats_logo" /></a>
 
 <!-- badges: start -->
 ![CRAN status](https://www.r-pkg.org/badges/version/biostats)
