### 2026-04-01 c4b579f
Update paper to reflect CRAN v.1.1.2 and GitHub pages
