### 2026-04-20 c5cb179
Updated Get started Vignette
