### 2026-04-20 #10317 @sebasquirarte
Thank you @rachelicr for your constructive feedback, I’m glad you liked the documentation and tutorials. I have reviewed and addressed all three issues raised and have made the necessary changes to the documentation and the GitHub Pages site. My response to each specific issue can be found in its corresponding thread.
