+++ pkgdown.yml
@@ -3,4 +3,4 @@ pkgdown: 2.2.0
 pkgdown_sha: ~
 articles:
   biostats: biostats.html
-last_built: 2026-04-20T17:52Z
+last_built: 2026-04-21T17:34Z
