+++ README.Rmd
@@ -64,6 +64,10 @@ pak::pak("sebasquirarte/biostats")
 library(biostats)
 ```
 
+## Documentation
+
+Online documentation for this package is available   [here](https://sebasquirarte.github.io/biostats/articles/biostats.html).
+
 ## Feedback
 
 We welcome feedback, suggestions, and bug reports. You can share your thoughts via email (sebastian.quirarte@sophia.com.mx) or [GitHub issues](https://github.com/sebasquirarte/biostats/issues). 
