+++ README.Rmd
@@ -55,7 +55,6 @@ install.packages("biostats")
 pak::pak("sebasquirarte/biostats")
 ```
 
-
 >**Note for Linux users**: Installing from source may require `cmake` to be installed
 > prior to package installation, as it is needed to build the `nloptr` dependency.
 
@@ -64,6 +63,10 @@ pak::pak("sebasquirarte/biostats")
 ```{r usage, warning=FALSE, message=FALSE, results='hide'}
 library(biostats)
 ```
+
+## Documentation
+
+Online documentation for this package is available   [here](https://sebasquirarte.github.io/biostats/articles/biostats.html).
 
 ## Feedback
 
