+++ README.md
@@ -58,6 +58,11 @@ pak::pak("sebasquirarte/biostats")
 library(biostats)
 ```
 
+## Documentation
+
+Online documentation for this package is available
+[here](https://sebasquirarte.github.io/biostats/articles/biostats.html).
+
 ## Feedback
 
 We welcome feedback, suggestions, and bug reports. You can share your
