### 2026-06-24 4969bdc
Minor fixes to paper based on JOSS editor feedback
