+++ paper.bib
@@ -127,7 +127,7 @@ without Propensity Score Weights},
 }
 
 @Manual{extras:2025,
-  title = {{e}xtras: Helper Functions for Bayesian Analyses},
+  title = {{e}xtras: Helper Functions for {B}ayesian Analyses},
   author = {Nicole Hill and Joe Thorley},
   year = {2025},
   note = {R package version 0.8.0},
@@ -148,7 +148,7 @@ Proportions in Phase III Clinical Trials},
 @Manual{TrialSize:2024,
   title = {TrialSize: R Functions for Chapter 3,4,6,7,9,10,11,12,14,15 of Sample Size
 Calculation in Clinical Research},
-  author = {Ed Zhang ; Vicky Qian Wu ; Shein-Chung Chow ; Harry G.Zhang},
+  author = {Ed Zhang and Vicky Qian Wu and Shein-Chung Chow and Harry G. Zhang},
   year = {2024},
   note = {R package version 1.4.1},
   url = {https://CRAN.R-project.org/package=TrialSize},
@@ -211,7 +211,7 @@ Calculation in Clinical Research},
   author    = {Chow, Shein-Chung and Shao, Jun and Wang, Hansheng and Lokhnygina, Yuliya},
   year      = {2017},
   edition   = {3},
-  publisher = {Chapman and Hall/CRC},
+  publisher = {Chapman & Hall/CRC},
   doi       = {10.1201/9781315183084}
 }
 
@@ -234,7 +234,3 @@ Calculation in Clinical Research},
   address   = {London},
   langid    = {en}
 }
-
-
-
-
