### 2026-06-24 fbbfa4d
Fix ggblanket citation on paper
