+++ paper.bib
@@ -108,13 +108,12 @@ without Propensity Score Weights},
   doi = {10.32614/cran.package.rstatix}
 }
 
-@Manual{ggblanket:2026,
-  title = {{g}gblanket: Publication-Quality 'ggplot2' Visualisation},
-  author = {David Hodge},
-  year = {2026},
-  note = {R package version 20.0.0},
-  url = {https://CRAN.R-project.org/package=ggblanket},
-  doi = {10.32614/CRAN.package.ggblanket}
+@Manual{ggblanket:2025,
+    title = {ggblanket: Simplify 'ggplot2' Visualisation},
+    author = {David Hodge},
+    year = {2025},
+    note = {R package version 12.4.0},
+    url = {https://CRAN.R-project.org/package=ggblanket},
 }
 
 @Manual{ggpubr:2025,
