### 2026-06-24 2590b7c
Citation minor fix
