+++ paper.bib
@@ -109,7 +109,7 @@ without Propensity Score Weights},
 }
 
 @Manual{ggblanket:2025,
-    title = {ggblanket: Simplify 'ggplot2' Visualisation},
+    title = {{g}gblanket: Simplify 'ggplot2' Visualisation},
     author = {David Hodge},
     year = {2025},
     note = {R package version 12.4.0},
