### 2026-06-24 ffcc664
Add CITATION.cff for software citation details
